DOOM-0055 (2/n, take 2): widen visplane to unsigned int per Crispy Doom recipe
My first widening (byte -> unsigned short, signed R_MakeSpans params) made
the floor render WORSE. Online research into how the canonical 2x hi-res
fork (Crispy Doom) solved this exact "byte top[]/bottom[] is the rub for
resolution change" problem found the two pieces I missed:

  1. visplane top[]/bottom[] (and their [minx-1]/[maxx+1] pads) must be
     unsigned int, with the unfilled-column sentinel 0xffffffffu (outside
     any row range; the memset 0xff byte-fill still yields it).
  2. R_MakeSpans' span params must be UNSIGNED. With signed int, the
     all-ones sentinel reads as -1 and inverts the span open/close logic,
     under-filling the floor -- exactly the regression I hit.

Also guard the sky draw loop to skip the sentinel column explicitly (the
byte version already mis-skipped sky at hi-res; make it robust).

The step-1 R_MapPlane guard stays as defense but should now be a no-op.
Builds clean. Pending user verify: floor fills cleanly (no smear), no menu
text left on the floor, sky correct, no crash -- and likely DOOM-0053
(menu ghosting) closes too since the view now fully repaints.

Sources: Crispy Doom r_defs.h / r_plane.c (fabiangreffrath/crispy-doom).

+++ linuxdoom-1.10/r_defs.h
@@ -466,16 +466,21 @@ typedef struct
   int			maxx;
   
   // leave pads for [minx-1]/[maxx+1]
-  
-  byte		pad1;
-  // Here lies the rub for all
-  //  dynamic resize/change of resolution.
-  byte		top[SCREENWIDTH];
-  byte		pad2;
-  byte		pad3;
-  // See above.
-  byte		bottom[SCREENWIDTH];
-  byte		pad4;
+
+  // "Here lies the rub for all dynamic resize/change of resolution": top[]/
+  // bottom[] hold physical screen row indices, but were byte (0..255), so at
+  // hi-res (SCREENHEIGHT=400) rows > 255 truncated and corrupted the spans
+  // (DOOM-0055). Widened to unsigned int per Crispy Doom's proven hi-res fix:
+  // the unfilled-column sentinel becomes 0xffffffffu (outside any row range)
+  // and R_MakeSpans takes UNSIGNED params so the sentinel reads as the largest
+  // value (a signed int would read all-ones as -1 and invert the span logic).
+  // The pads must match the element type so [minx-1]/[maxx+1] stay valid.
+  unsigned int	pad1;
+  unsigned int	top[SCREENWIDTH];
+  unsigned int	pad2;
+  unsigned int	pad3;
+  unsigned int	bottom[SCREENWIDTH];
+  unsigned int	pad4;
 
 } visplane_t;
